flask-server/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cohere==3.10.0
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.21.12
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
